refactor(tests): Pull out make_client function
